homework#22/requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
SQLAlchemy==1.4.39
Werkzeug==2.1.2